requirements.txt
coverage==4.5.3
django-celery-beat==1.4.0
django-compat==1.0.15
flower==0.9.3
ipython==7.5.0
pip-chill==0.1.8
psycopg2==2.8.2
redis==3.2.1
vk-api==11.4.0
